Getting a convex shape kinda works
There is an issue where when we are checking skybox-line intersections
with triangle line intersections. If the triangle is too big, the corner
points can't be projected onto the skybox-side since they fall outside
the plane.

Tried to fix it by orienting the plane-normal to coincide with the
view-direction while still letting the skybox-line lie on the plane.

This _seems_ to work for large triangles, but fail often for small
triangles and I don't know why.