Rework positioning of menu (Wip)
The positioning of the menu-items is being reworked to
be a bottom up-approach. That is a parent node gets its
size from its children. This mean that all leafs has to have a
ABSOLUTE_SIZE_ATTRIBTUE

Basically all containers work as grid and can stack its leafs either
vertically or horizontally.

Leaf nodes nned to have the ABSOLUTE_SIZE_ATTRIBUTE and
Root nodes need to have the POSITION_ATTRIBUTE.

While the previous implementation worked great for relative sizes which
was natural when designing borders, etc, it became cumbersome when doing
anything else. This bottom up approach will probably require a new
approach to borders and things that needs to have the size of the parent
node. Hopefully there will be a good solution for that.