Do root window border size with new scheme
We have completely disabled the tab-window.
The tab window sits in the root window and holds a TabGrid containing
tabs.
Tabs links to plugins.

This is a broken state of the gui. Pausing here to try and do a
immediate mode type gui instead.

If you wanna pick up here later, the next part should be to add back the
tab-window using the new Bottom up positioning system.